refactor(kas-draw): modularize main example
